requirements.txt
aiofiles==23.1.0
asgiref==3.6.0
blinker==1.6.2
boto3==1.25.5
certifi==2022.12.7
click==8.1.3
Deprecated==1.2.13
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
importlib-metadata==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.21.6
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-exporter-otlp-proto-grpc==1.17.0
opentelemetry-instrumentation==0.38b0
opentelemetry-instrumentation-asgi==0.38b0
opentelemetry-semantic-conventions==0.38b0
opentelemetry-util-http==0.38b0
pickle5==0.0.11
priority==2.0.0
protobuf==3.19.0
psutil==5.9.4
quart==0.18.3
scipy==1.7.3
sentry-sdk==1.16.0
sqlitedict==2.1.0
tensorflow==2.11.0
toml==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
wsproto==1.2.0
zipp==3.15.0

